+++ src/main/java/com/nexmo/client/application/CreateApplicationMethod.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2019 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,26 +19,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications;
+package com.nexmo.client.application;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.NexmoBadRequestException;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.TokenAuthMethod;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.BasicResponseHandler;
+import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-class ListApplicationsEndpoint extends AbstractMethod<ListApplicationsRequest, ListApplicationsResponse> {
-
+class CreateApplicationMethod extends AbstractMethod<Application, Application> {
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{TokenAuthMethod.class};
 
     private static final String PATH = "/applications";
 
-    ListApplicationsEndpoint(HttpWrapper httpWrapper) {
+    CreateApplicationMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
     }
 
@@ -48,15 +50,19 @@ class ListApplicationsEndpoint extends AbstractMethod<ListApplicationsRequest, L
     }
 
     @Override
-    public RequestBuilder makeRequest(ListApplicationsRequest request) throws NexmoClientException, UnsupportedEncodingException {
-        RequestBuilder requestBuilder = RequestBuilder.get(
-                httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH);
-        request.addParams(requestBuilder);
-        return requestBuilder;
+    public RequestBuilder makeRequest(Application application) throws NexmoClientException, UnsupportedEncodingException {
+        return RequestBuilder
+                .post(httpWrapper.getHttpConfig().getVersionedApiBaseUri("v2") + PATH)
+                .setHeader("Content-Type", "application/json")
+                .setEntity(new StringEntity(application.toJson()));
     }
 
     @Override
-    public ListApplicationsResponse parseResponse(HttpResponse response) throws IOException {
-        return ListApplicationsResponse.fromJson(new BasicResponseHandler().handleResponse(response));
+    public Application parseResponse(HttpResponse response) throws IOException, NexmoClientException {
+        if (response.getStatusLine().getStatusCode() != 201) {
+            throw new NexmoBadRequestException(EntityUtils.toString(response.getEntity()));
+        }
+
+        return Application.fromJson(new BasicResponseHandler().handleResponse(response));
     }
 }