+++ src/main/java/com/nexmo/client/application/GetApplicationMethod.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2019 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,28 +19,27 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications;
+package com.nexmo.client.application;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.NexmoBadRequestException;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.TokenAuthMethod;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.BasicResponseHandler;
+import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-class UpdateApplicationMethod extends AbstractMethod<UpdateApplicationRequest, ApplicationDetails> {
-
+class GetApplicationMethod extends AbstractMethod<String, Application> {
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{TokenAuthMethod.class};
 
-    private static final String PATH = "/applications/";
+    private static final String PATH = "/applications/%s";
 
-    UpdateApplicationMethod(HttpWrapper httpWrapper) {
+    GetApplicationMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
     }
 
@@ -50,20 +49,18 @@ class UpdateApplicationMethod extends AbstractMethod<UpdateApplicationRequest, A
     }
 
     @Override
-    public RequestBuilder makeRequest(UpdateApplicationRequest request) throws NexmoClientException, UnsupportedEncodingException {
-        return RequestBuilder.put(
-                httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH + request.getApplicationId())
-                .setHeader("Content-Type", "application/json")
-                .setEntity(new StringEntity(request.toJson(), ContentType.APPLICATION_JSON));
+    public RequestBuilder makeRequest(String id) throws NexmoClientException, UnsupportedEncodingException {
+        return RequestBuilder
+                .get(httpWrapper.getHttpConfig().getVersionedApiBaseUri("v2") + String.format(PATH, id))
+                .setHeader("Content-Type", "application/json");
     }
 
     @Override
-    public ApplicationDetails parseResponse(HttpResponse response) throws IOException {
-        return ApplicationDetails.fromJson(new BasicResponseHandler().handleResponse(response));
-    }
+    public Application parseResponse(HttpResponse response) throws IOException, NexmoClientException {
+        if (response.getStatusLine().getStatusCode() != 200) {
+            throw new NexmoBadRequestException(EntityUtils.toString(response.getEntity()));
+        }
 
-    @Override
-    protected RequestBuilder applyAuth(RequestBuilder request) throws NexmoClientException {
-        return getAuthMethod(getAcceptableAuthMethods()).applyAsJsonProperties(request);
+        return Application.fromJson(new BasicResponseHandler().handleResponse(response));
     }
 }