+++ src/main/java/com/nexmo/client/application/UpdateApplicationMethod.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 Nexmo Inc
+ * Copyright (c) 2011-2019 Nexmo Inc
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,26 +19,28 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.applications;
+package com.nexmo.client.application;
 
 import com.nexmo.client.AbstractMethod;
 import com.nexmo.client.HttpWrapper;
+import com.nexmo.client.NexmoBadRequestException;
 import com.nexmo.client.NexmoClientException;
 import com.nexmo.client.auth.TokenAuthMethod;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.BasicResponseHandler;
+import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 
-class DeleteApplicationMethod extends AbstractMethod<String, Void> {
-
+class UpdateApplicationMethod extends AbstractMethod<Application, Application> {
     private static final Class[] ALLOWED_AUTH_METHODS = new Class[]{TokenAuthMethod.class};
 
-    private static final String PATH = "/applications/";
+    private static final String PATH = "/applications/%s";
 
-    DeleteApplicationMethod(HttpWrapper httpWrapper) {
+    UpdateApplicationMethod(HttpWrapper httpWrapper) {
         super(httpWrapper);
     }
 
@@ -48,13 +50,19 @@ class DeleteApplicationMethod extends AbstractMethod<String, Void> {
     }
 
     @Override
-    public RequestBuilder makeRequest(String applicationId) throws NexmoClientException, UnsupportedEncodingException {
-        return RequestBuilder.delete(httpWrapper.getHttpConfig().getVersionedApiBaseUri("v1") + PATH + applicationId);
+    public RequestBuilder makeRequest(Application application) throws NexmoClientException, UnsupportedEncodingException {
+        return RequestBuilder
+                .put(httpWrapper.getHttpConfig().getVersionedApiBaseUri("v2") + String.format(PATH, application.getId()))
+                .setHeader("Content-Type", "application/json")
+                .setEntity(new StringEntity(application.toJson()));
     }
 
     @Override
-    public Void parseResponse(HttpResponse response) throws IOException {
-        new BasicResponseHandler().handleResponse(response);
-        return null;
+    public Application parseResponse(HttpResponse response) throws IOException, NexmoClientException {
+        if (response.getStatusLine().getStatusCode() != 200) {
+            throw new NexmoBadRequestException(EntityUtils.toString(response.getEntity()));
+        }
+
+        return Application.fromJson(new BasicResponseHandler().handleResponse(response));
     }
 }